+++ console2/src/components/organisms/EncryptValueActivity/index.tsx
@@ -140,7 +140,9 @@ class EncryptValueActivity extends React.PureComponent<ExternalProps, State> {
                                             <Icon
                                                 name="copy"
                                                 link={true}
-                                                onClick={() => (copyToClipboard as any)(result.data)}
+                                                onClick={() =>
+                                                    (copyToClipboard as any)(result.data)
+                                                }
                                             />
                                         }
                                         fluid={true}