+++ console2/src/components/organisms/ProjectListActivity/index.tsx
@@ -23,20 +23,39 @@ import { Input, Menu } from 'semantic-ui-react';
 
 import { ConcordKey } from '../../../api/common';
 import { ProjectList, RedirectButton } from '../../organisms';
+import { connect, DispatchProp } from 'react-redux';
+import { State as SessionState } from '../../../state/session';
+import { Organizations } from '../../../state/data/orgs/types';
 
 interface State {
     filter?: string;
 }
 
-interface Props {
+interface ExternalProps {
     orgName: ConcordKey;
 }
+
+interface UserProps {
+    orgs: Organizations;
+}
+
+type Props = ExternalProps & UserProps & DispatchProp<any>;
 
 class ProjectListActivity extends React.Component<Props, State> {
     constructor(props: Props) {
         super(props);
         this.state = {};
     }
+
+    isUserOrgMember = (orgName: string) => {
+        const userOrgs = this.props.orgs;
+
+        return (
+            Object.keys(userOrgs)
+                .map((k) => userOrgs[k])
+                .filter((org) => org.name === orgName).length > 0
+        );
+    };
 
     render() {
         const { orgName } = this.props;
@@ -54,6 +73,7 @@ class ProjectListActivity extends React.Component<Props, State> {
 
                     <Menu.Item position={'right'}>
                         <RedirectButton
+                            disabled={!this.isUserOrgMember(orgName)}
                             icon="plus"
                             positive={true}
                             labelPosition="left"
@@ -68,5 +88,7 @@ class ProjectListActivity extends React.Component<Props, State> {
         );
     }
 }
-
-export default ProjectListActivity;
+const mapStateToProps = ({ session }: { session: SessionState }): UserProps => ({
+    orgs: session.user.orgs ? session.user.orgs : {}
+});
+export default connect(mapStateToProps)(ProjectListActivity);